favorite video games database migration room

+++ core/core-database-domain/src/main/java/com/example/core_database_domain/userCase/favoriteVideoGame/GetFavoriteVideoGamesCountUseCase.kt
@@ -7,5 +7,5 @@ import javax.inject.Inject
 class GetFavoriteVideoGamesCountUseCase @Inject constructor(
     private val repository: FavoriteVideoGameRepository
 ) {
-    operator fun invoke(): Flow<Long> = repository.getCount()
+    operator fun invoke(): Flow<Int> = repository.getCount()
 }
